WIP first draft part 2
these commits are terrible, but i had two days to throw this together
and was working to create a new TypeScript IDE between sublime and my
terminal

package.json declares a couple of scripts for tdd that use a compile
step and a test step for individual running.  requirements included.

got nearly 40% through writing tests, but even that was 1400 lines.