add travis build configuration to project
moved to .nvmrc for versioning - travis supports up to 4.1 so that's
where we'll start.

locked chai and mocha dependencies since the versions supplied in the
DefinitelyTyped definitions are restricted there.  nothing broke with
the lock, but i probably should go back through the chai and mocha
changelogs to see what i'm missing out on by using these older versions.